refactor: move main app component to pages directory
