+++ src/main.tsx
@@ -1,5 +1,5 @@
 import { createRoot } from 'react-dom/client';
-import App from './App.tsx';
+import App from './pages/App.tsx';
 import './styles/main.scss';
 
 createRoot(document.getElementById('root')!).render(<App />);
