+++ src/pages/App.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
-import { generateItems, generateLives, generateShells, usePhone } from "./utils/gameUtils";
-import Button from "./components/Button";
-import Modal from "./components/Modal";
-import generateRandomID from "./utils/generateRandomID";
+import { generateItems, generateLives, generateShells, usePhone } from "../utils/gameUtils";
+import Button from "../components/Button";
+import Modal from "../components/Modal";
+import generateRandomID from "../utils/generateRandomID";
 
 export type Shell = {
     type: string,
